+++ dds_cli/__main__.py
@@ -42,6 +42,7 @@ from dds_cli.options import (
     sort_projects_option,
     source_option,
     source_path_file_option,
+    token_path_option,
     break_on_fail_flag,
     json_flag,
     nomail_flag,
@@ -72,8 +73,6 @@ click.rich_click.MAX_WIDTH = 100
 #                                                                                                  #
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # ##
 
-# Get token metadata
-username = dds_cli.user.User.get_user_name_if_logged_in()
 
 # Print header to STDERR
 dds_cli.utils.stderr_console.print(
@@ -83,7 +82,6 @@ dds_cli.utils.stderr_console.print(
     "\n[green] ︶  (  ) ) ([/]    [blue][link={0}]{0}[/link]".format(dds_cli.__url__),
     f"\n[green]      ︶ (  )[/]    [dim]Version {dds_cli.__version__}",
     "\n[green]          ︶",
-    f"\n[green]Current user:[/] [red]{username}" if username else "",
     highlight=False,
 )
 
@@ -96,13 +94,22 @@ dds_cli.utils.stderr_console.print(
 @click.option(
     "--no-prompt", is_flag=True, default=False, help="Run without any interactive features."
 )
+@token_path_option()
 @click.version_option(version=dds_cli.__version__, prog_name=dds_cli.__title__)
 @click.pass_context
-def dds_main(click_ctx, verbose, log_file, no_prompt):
+def dds_main(click_ctx, verbose, log_file, no_prompt, token_path):
     """SciLifeLab Data Delivery System (DDS) command line interface.
 
     Access token is saved in a .dds_cli_token file in the home directory.
     """
+    # Get token metadata
+    username = dds_cli.user.User.get_user_name_if_logged_in(token_path=token_path)
+
+    if username:
+        dds_cli.utils.stderr_console.print(
+            f"[green]Current user:[/] [red]{username}", highlight=False
+        )
+
     if "--help" not in sys.argv:
         # Set the base logger to output DEBUG
         LOG.setLevel(logging.DEBUG)
@@ -128,7 +135,7 @@ def dds_main(click_ctx, verbose, log_file, no_prompt):
             LOG.addHandler(log_fh)
 
         # Create context object
-        click_ctx.obj = {"NO_PROMPT": no_prompt}
+        click_ctx.obj = {"NO_PROMPT": no_prompt, "TOKEN_PATH": token_path}
 
 
 # ************************************************************************************************ #
@@ -167,6 +174,7 @@ def list_projects_and_contents(
                 show_usage=usage,
                 no_prompt=click_ctx.get("NO_PROMPT", False),
                 json=json,
+                token_path=click_ctx.get("TOKEN_PATH"),
             ) as lister:
                 projects = lister.list_projects(sort_by=sort)
                 if json:
@@ -201,6 +209,7 @@ def list_projects_and_contents(
                 tree=tree,
                 no_prompt=click_ctx.get("NO_PROMPT", False),
                 json=json,
+                token_path=click_ctx.get("TOKEN_PATH"),
             ) as lister:
                 if json:
                     json_output = {"project_name": project}
@@ -311,7 +320,7 @@ def login(click_ctx):
     if no_prompt:
         LOG.warning("The --no-prompt flag is ignored for `dds auth login`")
     try:
-        with dds_cli.auth.Auth():
+        with dds_cli.auth.Auth(token_path=click_ctx.get("TOKEN_PATH")):
             # Authentication token renewed in the init method.
             LOG.info("[green] :white_check_mark: Authentication token renewed![/green]")
     except (
@@ -326,13 +335,16 @@ def login(click_ctx):
 
 # -- dds auth logout -- #
 @auth_group_command.command(name="logout")
-def logout():
+@click.pass_obj
+def logout(click_ctx):
     """End authenticated session.
 
     Removes the saved authentication token by deleting the '.dds_cli_token' file.
     """
     try:
-        with dds_cli.auth.Auth(authenticate=False) as authenticator:
+        with dds_cli.auth.Auth(
+            authenticate=False, token_path=click_ctx.get("TOKEN_PATH")
+        ) as authenticator:
             authenticator.logout()
 
     except dds_cli.exceptions.DDSCLIException as err:
@@ -342,7 +354,8 @@ def logout():
 
 # -- dds auth info -- #
 @auth_group_command.command(name="info")
-def info():
+@click.pass_obj
+def info(click_ctx):
     """Display information about token.
 
     Information displayed:
@@ -351,7 +364,9 @@ def info():
     - Time of token expiration
     """
     try:
-        with dds_cli.auth.Auth(authenticate=False) as authenticator:
+        with dds_cli.auth.Auth(
+            authenticate=False, token_path=click_ctx.get("TOKEN_PATH")
+        ) as authenticator:
             authenticator.check()
     except dds_cli.exceptions.DDSCLIException as err:
         LOG.error(err)
@@ -412,7 +427,8 @@ def add_user(click_ctx, email, role, project, unit, no_mail):
     """
     try:
         with dds_cli.account_manager.AccountManager(
-            no_prompt=click_ctx.get("NO_PROMPT", False)
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as inviter:
             inviter.add_user(email=email, role=role, project=project, no_mail=no_mail, unit=unit)
     except (
@@ -469,6 +485,7 @@ def delete_user(click_ctx, email, self):
             with dds_cli.account_manager.AccountManager(
                 method="delete",
                 no_prompt=click_ctx.get("NO_PROMPT", False),
+                token_path=click_ctx.get("TOKEN_PATH"),
             ) as manager:
                 if self and not email:
                     manager.delete_own_account()
@@ -500,7 +517,8 @@ def get_info_user(click_ctx):
     """Display information connected to your own DDS account."""
     try:
         with dds_cli.account_manager.AccountManager(
-            no_prompt=click_ctx.get("NO_PROMPT", False)
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as get_info:
             get_info.get_user_info()
     except (
@@ -537,6 +555,7 @@ def activate_user(click_ctx, email):
         try:
             with dds_cli.account_manager.AccountManager(
                 no_prompt=click_ctx.get("NO_PROMPT", False),
+                token_path=click_ctx.get("TOKEN_PATH"),
             ) as manager:
                 manager.user_activation(email=email, action="reactivate")
 
@@ -574,6 +593,7 @@ def deactivate_user(click_ctx, email):
         try:
             with dds_cli.account_manager.AccountManager(
                 no_prompt=click_ctx.get("NO_PROMPT", False),
+                token_path=click_ctx.get("TOKEN_PATH"),
             ) as manager:
                 manager.user_activation(email=email, action="deactivate")
 
@@ -680,7 +700,8 @@ def create(
     """Create a project."""
     try:
         with dds_cli.project_creator.ProjectCreator(
-            no_prompt=click_ctx.get("NO_PROMPT", False)
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as creator:
             emails_roles = []
             if owner or researcher:
@@ -748,7 +769,9 @@ def display_project_status(click_ctx, project, show_history):
     """Display and manage project statuses."""
     try:
         with dds_cli.project_status.ProjectStatusManager(
-            project=project, no_prompt=click_ctx.get("NO_PROMPT", False)
+            project=project,
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as updater:
             updater.get_status(show_history)
     except (
@@ -777,7 +800,9 @@ def release_project(click_ctx, project, deadline, no_mail):
     """Make project data available for user download."""
     try:
         with dds_cli.project_status.ProjectStatusManager(
-            project=project, no_prompt=click_ctx.get("NO_PROMPT", False)
+            project=project,
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as updater:
             updater.update_status(new_status="Available", deadline=deadline, no_mail=no_mail)
     except (
@@ -802,7 +827,9 @@ def retract_project(click_ctx, project):
     """
     try:
         with dds_cli.project_status.ProjectStatusManager(
-            project=project, no_prompt=click_ctx.get("NO_PROMPT", False)
+            project=project,
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as updater:
             updater.update_status(new_status="In Progress")
     except (
@@ -827,7 +854,9 @@ def archive_project(click_ctx, project):
     """
     try:
         with dds_cli.project_status.ProjectStatusManager(
-            project=project, no_prompt=click_ctx.get("NO_PROMPT", False)
+            project=project,
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as updater:
             updater.update_status(new_status="Archived")
     except (
@@ -852,7 +881,9 @@ def delete_project(click_ctx, project):
     """
     try:
         with dds_cli.project_status.ProjectStatusManager(
-            project=project, no_prompt=click_ctx.get("NO_PROMPT", False)
+            project=project,
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as updater:
             updater.update_status(new_status="Deleted")
     except (
@@ -877,7 +908,9 @@ def abort_project(click_ctx, project):
     """
     try:
         with dds_cli.project_status.ProjectStatusManager(
-            project=project, no_prompt=click_ctx.get("NO_PROMPT", False)
+            project=project,
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as updater:
             updater.update_status(new_status="Archived", is_aborted=True)
     except (
@@ -917,7 +950,8 @@ def grant_project_access(click_ctx, project, email, owner, no_mail):
     """Grant user access to a project."""
     try:
         with dds_cli.account_manager.AccountManager(
-            no_prompt=click_ctx.get("NO_PROMPT", False)
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as granter:
             role = "Researcher"
             if owner:
@@ -943,7 +977,8 @@ def revoke_project_access(click_ctx, project, email):
     """Revoke user access to a project."""
     try:
         with dds_cli.account_manager.AccountManager(
-            no_prompt=click_ctx.get("NO_PROMPT", False)
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as revoker:
             revoker.revoke_project_access(project, email)
     except (
@@ -967,7 +1002,8 @@ def fix_project_access(click_ctx, email, project):
     """Re-grant project access to user that has lost access due to password reset."""
     try:
         with dds_cli.account_manager.AccountManager(
-            no_prompt=click_ctx.get("NO_PROMPT", False)
+            no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as fixer:
             fixer.fix_project_access(email=email, project=project)
     except (
@@ -1045,6 +1081,7 @@ def put_data(
             num_threads=num_threads,
             silent=silent,
             no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         )
     except (
         dds_cli.exceptions.AuthenticationError,
@@ -1125,6 +1162,7 @@ def get_data(
             silent=silent,
             verify_checksum=verify_checksum,
             no_prompt=click_ctx.get("NO_PROMPT", False),
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as getter:
 
             with rich.progress.Progress(
@@ -1296,6 +1334,7 @@ def rm_data(click_ctx, project, file, folder, rm_all):
         with dds_cli.data_remover.DataRemover(
             project=project,
             no_prompt=no_prompt,
+            token_path=click_ctx.get("TOKEN_PATH"),
         ) as remover:
 
             if rm_all: