+++ dds_cli/data_getter.py
@@ -50,6 +50,7 @@ class DataGetter(base.DDSBaseClass):
         verify_checksum: bool = False,
         method: str = "get",
         no_prompt: bool = False,
+        token_path: str = None,
     ):
         """Handle actions regarding downloading data."""
         # Initiate DDSBaseClass to authenticate user
@@ -58,6 +59,7 @@ class DataGetter(base.DDSBaseClass):
             dds_directory=destination,
             method=method,
             no_prompt=no_prompt,
+            token_path=token_path,
         )
 
         # Initiate DataGetter specific attributes